% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{canny}
\alias{canny}
\title{Apply the Canny edge detection}
\usage{
canny(im, t1, t2, alpha = 1, sigma = 2, smoothing = 0)
}
\arguments{
\item{im}{an image.}

\item{t1, }{threshold for weak edges (if missing, both thresholds are determined automatically).}

\item{t2}{threshold for strong edges.}

\item{alpha}{threshold adjustment factor (default 1).}

\item{sigma}{degree of smoothing image gradient.}

\item{smoothing}{numeric (integer). Smoothness of image texture.}
}
\value{
an array of the edge image.
}
\description{
This is a wrapper of cannyEdges() function in the imager package.
t1 and t2 determines the edge threshold.
if the threshold parameters are missing, they are determined automatically using a k-means heuristic.
alpha parameter adjusts the automatic thresholds up or down.
The edge detection is based on a smoothed image gradient (set by the sigma parameter).
If smoothing parameter is set, high frequency noise is removed before applying the Canny edge detection.
}
\examples{
im = canny(face)
pplot(im)
}
