% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{face}
\alias{face}
\title{A face image.}
\format{
An array with 600 x 460 * 3 dimensions.
Each dimension represents y-coordinate, x-coordinate, and color channel.
}
\usage{
face
}
\description{
A photograph obtained from a free stock photos site.
pexels.com/photo/man-about-to-touch-his-face-wearing-blue-suit-718261/
}
\examples{
plot(face)
}
\keyword{datasets}
