% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{im_gray}
\alias{im_gray}
\title{Convert to grayscale}
\usage{
im_gray(im, drop = TRUE)
}
\arguments{
\item{im}{an image (array).}

\item{drop}{if TRUE (default), returns an image with a single channel,
otherwise keep the three channels.}
}
\value{
an array of the grayscaled image
}
\description{
Convert a color image to grayscale.
}
\examples{
pplot(face) # color
im = im_gray(face)
pplot(im) # grayscale
}
