% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{im_save}
\alias{im_save}
\title{Save an image to disk}
\usage{
im_save(im, name, path = getwd(), format = "png", quality = 0.95)
}
\arguments{
\item{im}{An image.}

\item{name}{Name of the image file.}

\item{path}{Path to file.}

\item{format}{Image format. Either "jpg", "png", "tiff", or "bmp". Default is "png".}

\item{quality}{(jpg only) default is 0.95. Higher quality means less compression.}
}
\value{
No return value, called for side effects.
}
\description{
Save an image to disk
}
\examples{
\dontrun{
imsave( face )
imsave( face, name = "myimage", path = "path/to/image", format = "jpg" )
}
}
