% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{im_scale}
\alias{im_scale}
\title{Resize image}
\usage{
im_scale(im, scale = 1)
}
\arguments{
\item{im}{an image (array).}

\item{scale}{scale factor}
}
\value{
an array of the scaled image.
}
\description{
Resize image by a single scale factor.
}
\examples{
dim(face) # original size
im = im_scale(face, 0.5)
dim(im) # resized to half of its original size
}
