% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.nimg}
\alias{plot.nimg}
\title{Display an image}
\usage{
\method{plot}{nimg}(x, rescale = FALSE, ...)
}
\arguments{
\item{x}{an image}

\item{rescale}{logical. if true, then pixel value is rescaled to range between 0 and 1.}

\item{...}{other parameters to be passed to plot.default}
}
\value{
No return value, called for side effects.
}
\description{
Display an image
}
\examples{
plot(face)
}
