% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pplot}
\alias{pplot}
\title{Plot an image}
\usage{
pplot(im, rescale = FALSE)
}
\arguments{
\item{im}{an image (array).}

\item{rescale}{logical. if TRUE, pixel values are rescaled so that they are in [0,1].}
}
\value{
No return value, called for side effects.
}
\description{
Plot an image
}
\examples{
pplot(face)
}
