% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{face}
\alias{face}
\title{A face image.}
\format{
An array with 800 * 800 * 3 * 1 dimensions.
Each dimension represents y-coordinate, x-coordinate, color channel, and frame.
}
\usage{
face
}
\description{
A photograph obtained from a free stock photos site.
pexels.com/photo/woman-wearing-sunhat-1382731/
}
\examples{
pplot(face)
}
\keyword{datasets}
