% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{im_load}
\alias{im_load}
\title{Load image from file}
\usage{
im_load(path)
}
\arguments{
\item{path}{path to an image file}
}
\value{
an array representing the image
}
\description{
The png, jpg, and bmp are supported.
}
\examples{
\dontrun{
im = im_load("path/to/your/image.jpg")
pplot(im)
}
}
