% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sketch}
\alias{sketch}
\title{Apply a sketch effect on an image}
\usage{
sketch(
  im,
  style = 1,
  lineweight = 1,
  smoothing = 1,
  contrast = ifelse(style == 1, 8, 3),
  gain = 0.1
)
}
\arguments{
\item{im}{an image (array).}

\item{style}{numeric (integer). Either 1 (edge-focused) or 2 (shading preserving)}

\item{lineweight}{numeric (integer). Strength of lines.}

\item{smoothing}{numeric (integer). Smoothness of image texture.}

\item{contrast}{numeric (integer). Adjusts the image contrast.}

\item{gain}{numeric (double). Can be used to reduce noise in dim regions.}
}
\value{
an array of the sketched image.
}
\description{
Apply a sketch effect on an image
}
\examples{
im = sketch(face, style = 1, lineweight = 1, smoothing = 1)
pplot(im)

\dontrun{
im = im_load("path/to/your/image.jpg")
pplot(im)
}
}
