% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{survey}
\alias{survey}
\title{Create multiple sketches at once and combine them into a single image}
\usage{
survey(
  im,
  style = 1,
  weight_levels = c(1, 3, 5),
  smooth_levels = c(0, 2, 4),
  contrast,
  verbose = TRUE
)
}
\arguments{
\item{im}{an image.}

\item{style}{numeric (integer). Either 1 (edge-focused) or 2 (smooth gradient)}

\item{weight_levels}{numeric (integer). a vector of lineweight values}

\item{smooth_levels}{numeric (integer). a vector of smoothing values}

\item{contrast}{numeric (integer). Adjusts the image contrast.}

\item{verbose}{If TRUE (default), progress information is displayed in the Console.}
}
\value{
an array of the sketched image.
}
\description{
It is often necessary to find optimal sketch style parameters for your task.
With this function, you can easily compare the effects of different style parameters.
}
\examples{
\donttest{
im = survey(face)
pplot(im)

im = survey(face, style = 1, weight_levels = c(1, 2), smooth_levels = c(0, 2), contrast = 9)
pplot(im)
}
}
